First fix - from B1397KB
The translate.py file is temporarily removed, and i18n internationalization is added after the project is basically completed. The format of the bot.py file is optimized, and the function definition is placed in the function.py file to reduce the size of the bot.py file.

Modify the platoon command, add the command to obtain the BFBAN robot account token, and add captcha verification code image conversion.